Move some helpers to `magento-utils.js` as per PR feedback
